feat: move wallet provider verification to context
